Migrated AttributeTypeEditor to new Type Editing Infrastructure
The existing AttributeTypeEditor was migrated to an ITypeEditorPage and
a new shell AttributeTypeEditor is introduced. This is the basis for
adding the DescriptionEditor later.
