+++ plugins/org.eclipse.fordiac.ide.attributetypeeditor/src/org/eclipse/fordiac/ide/attributetypeeditor/Messages.java
@@ -18,18 +18,11 @@ import org.eclipse.osgi.util.NLS;
 @SuppressWarnings("squid:S3008") // tell sonar the java naming convention does not make sense for this class
 public final class Messages extends NLS {
 	private static final String BUNDLE_NAME = "org.eclipse.fordiac.ide.attributetypeeditor.messages"; //$NON-NLS-1$
-	public static String ErrorCompositeMessage;
-	public static String StructViewingComposite_Headline;
-	public static String MessageDialogTitle_OutsideWorkspaceError;
-	public static String MessageDialogContent_OutsideWorkspaceError;
-	public static String Editor_ComboBox_StructEditingComposite;
-	public static String Editor_ComboBox_DirectlyDerivedTypeComposite;
-	public static String MessageDialogTitle_SaveError;
-	public static String MessageDialogContent_SaveError;
-	public static String StructAlteringButton_SaveAndUpdate;
-	public static String StructAlteringButton_SaveAs;
-	public static String StructAlteringDescription;
-	public static String WarningDialog_StructNotSaved;
+
+	public static String AttributeTypeEditorPage_Title;
+
+	public static String AttributeTypeEditorPage_ComboBox_StructEditingComposite;
+	public static String AttributeTypeEditorPage_ComboBox_DirectlyDerivedTypeComposite;
 
 	static {
 		// initialize resource bundle
