+++ plugins/org.eclipse.fordiac.ide.attributetypeeditor/src/org/eclipse/fordiac/ide/attributetypeeditor/messages.properties
@@ -1,5 +1,5 @@
  ################################################################################
- # Copyright (c) 2023 Johannes Kepler University, Linz
+ # Copyright (c) 2023, 2024 Johannes Kepler University, Linz
  #
  # This program and the accompanying materials are made available under the
  # terms of the Eclipse Public License 2.0 which is available at
@@ -11,17 +11,8 @@
  #   Sebastian Hollersbacher
  #     - initial API and implementation and/or initial documentation
  ################################################################################
-ErrorCompositeMessage=Loading has failed
-StructViewingComposite_Headline=ATTRIBUTE Editor
-MessageDialogTitle_OutsideWorkspaceError=Not in workspace
-MessageDialogContent_OutsideWorkspaceError=Files outside of workspace cannot be opened
-Editor_ComboBox_StructEditingComposite=StructType
-Editor_ComboBox_DirectlyDerivedTypeComposite=DirectlyDerivedType
-MessageDialogTitle_SaveError=Saving not possible
-MessageDialogContent_SaveError=An error has occurred while saving
-ErrorCompositeMessage=Loading has failed
-StructViewingComposite_Headline=STRUCT Editor
-StructAlteringButton_SaveAndUpdate=Save and Update
-StructAlteringButton_SaveAs=Save As
-StructAlteringDescription=The following blocks were affected by changing this struct:
-WarningDialog_StructNotSaved=The changes have not been saved!
+
+AttributeTypeEditorPage_Title=Attribute declaration
+
+AttributeTypeEditorPage_ComboBox_StructEditingComposite=Struct type
+AttributeTypeEditorPage_ComboBox_DirectlyDerivedTypeComposite=Directly derived type
